修改部分字段类型，重新生成数据库代码

+++ src/main/java/com/seekerhut/mapper/ForumSectionUserRelationDAO.java
@@ -7,5 +7,5 @@ import org.springframework.stereotype.Repository;
  * ForumSectionUserRelationDAO继承基类
  */
 @Repository
-public interface ForumSectionUserRelationDAO extends MyBatisBaseDao<ForumSectionUserRelation, Integer> {
+public interface ForumSectionUserRelationDAO extends MyBatisBaseDao<ForumSectionUserRelation, Long> {
 }
